Created a more cache efficient version of Floyd Warshall by performing vertical matrix traversal in the inner loop instead of horizontal.
Also dist(u, pivot) are checked in the outer-loop to avoid unnecessary iterations.

Benchmarks:
|V| = 2000

Complete Graph
New :    10.364 s (5 allocations: 61.04 MiB)
Old :    158.904 s (5 allocations: 61.04 MiB)

Path Digraph
New :    75.381 ms (5 allocations: 61.04 MiB)
Old : 129.149 s (5 allocations: 61.04 MiB)

The speed up in the Path Digraph is due to avoiding unnecessary iterations.